Improve handling of invalid Relay ids in profile related objects removal
The email, phone and address objects of a profile are identified by a
global id when removed. The provided id can be incorrect in many ways.
It can for example be an id for a wrong kind of object or can point to a
non existing object. For all kinds of incorrect ids, produce an "object
not found" error. This doesn't reveal to the caller what exactly was the
specific reason so that it won't be possible to extract any
unintentional details from the state of the database.