Separate profile data update from persistance
Enable users of the profile updater helpers to do something in between
the object update (in memory) and persistance to database. None of the
users of the helpers in this commit yet use this possibility; they just
use the new update_profile_to_db helper, which does exactly the same
thing what the previous update_profile did.