HP-1710 Add management command for setting allowed data fields
The command reads the specification in JSON format from stdin and then
passes the data to the allowed data fields configurator function.

If nothing gets inserted into the stdin of the command, it would wait
forever. Pressing Ctrl+C interrupts the command, raising a
KeyboardInterrupt exception. That exception is "swallowed" and the
command just exits normally in that case.