Treat seed_data management command's data as development data
The real allowed data fields and services data does not live in this
repository. All the data that this management command creates should be
treated as development data. Thus renamed the command to
seed_development_data and removed the --development option.

The command does not get executed in container startup at all unless the
SEED_DEVELOPMENT_DATA environment variable is set.