Remove the `clear` option from seed_development_data management command
Having both --clear and --no-clear options is a bit useless. Clearing is
now the default and it can be prevented with the --no-clear option.

The --no-clear option is required in tests. Combined two very similar
tests to parameterized tests as the only difference was whether they
added a superuser or not.