feat: shorten profile name fields to 150 characters
First and last name need to be limited to 150 characters on the Profile model.
Updates to name information are also sent to keycloak, which will put the name
information into the API tokens that it produces.

User model has 150 character name fields. Upon receiving an API token
this backend will try to update the user object with the information in
the token. If the name is over 150 characters an error will be produced
and the user will be unable to use the API. Users with this error will
also be unable to resolve the issue by themselves.

Refs: HP-2344